requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
deprecation==2.1.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multidict==6.2.0
narwhals==1.33.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
postgrest==0.19.3
propcache==0.3.0
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
realtime==2.4.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.11.3
streamlit==1.44.1
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websockets==14.2
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
Flask==3.1.0
flask-cors==5.0.1
